pri_prog_ma/requirements.txt
libs/dq_ems-1.0.0-py3-none-any.whl
libs/deepocean_trade_api-3.4.0-py3-none-any.whl
google-cloud-firestore
google-cloud-logging
google-cloud-secret-manager
google-cloud-pubsub
yaml-1.3==0.1.0
Flask==2.0.2
gunicorn==20.1.0
ccxt
mockito
pytest
pytest-mockito
PyYAML==6.0
pytz
redis
structlog
influxdb
pandas
settrade==0.5.0
settrade-v2
fastapi
uvicorn
timeout-decorator
psutil
sendgrid==6.9.7
twilio==7.8.0
openpyxl
celery
python-dateutil
